fixes api translator/coordinator

+++ pages/api/projects/[code]/translators/[id].js
@@ -6,17 +6,27 @@ export default async function languageProjectTranslatorHandler(req, res) {
   }
   supabase.auth.setAuth(req.headers.token)
   const {
-    body: { project_id },
-    query: { id },
+    query: { code, id },
     method,
   } = req
+
   switch (method) {
     case 'DELETE':
+      const { data: project, error } = await supabase
+        .from('projects')
+        .select('id, code')
+        .eq('code', code)
+        .limit(1)
+        .maybeSingle()
+      if (error) throw error
+      if (!project?.id) {
+        return
+      }
       try {
         const { data, error } = await supabase
           .from('project_translators')
           .delete()
-          .match({ project_id, user_id: id })
+          .match({ project_id: project.id, user_id: id })
 
         if (error) throw error
         res.status(200).json(data)
